feat: reduce the 'c' variable to 8 bit
The restriction of the variable 'c' to 8 bits is needed to watch the
variable roll round to zero while incrementing beyond the maximal
unsigned 8-bit-number (255).

Instructions:
Run the debug in simulation mode and watch the a and c variables with
the logic analyzer. Hint: Press the 'Min/Max' button.

+++ main.c
@@ -10,7 +10,7 @@
 uint16_t a = 0;
 uint16_t range = 300;
 uint16_t locked = 0;
-uint16_t c = 0;
+uint8_t c = 0;
 uint32_t xPSR = 0;
 
 int main (void)
